pdf-to-markdown/pyproject.toml
[project]
name = "pdf-to-markdown-api"
version = "1.0.0"
description = "AI-powered PDF to Markdown conversion API with OpenAI-compatible interface"
authors = [{name = "Shili Intelligence"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0"}

dependencies = [
    "marker-pdf>=1.10.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
pdf2md-api = "scripts.api_server:main"
pdf2md-convert = "scripts.convert:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
